\name{geospt-package}
\alias{geospt-package}
\alias{geospt}
\docType{package}
\title{
Geostatistical Analysis and Design of Optimal Spatial Sampling Networks
}
\description{
A set of functions for: estimation of the variogram through trimmed mean, radial basis functions 
(optimization, prediction and cross-validation), summary statistics from cross-validation, pocket plot, and
design of optimal sampling networks through sequential and simultaneous points methods}
\details{
\tabular{ll}{
Package: \tab geospt\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2025-03-07\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Carlos Melo <cmelo@udistrital.edu.co>, Ali Santacruz, Oscar Melo <oomelom@unal.edu.co

Maintainer: Ali Santacruz <amsantac@unal.edu.co>
}
\keyword{ package }
\keyword{ spatial }
\seealso{
\code{\link{rbf}}, \code{\link{est.variograms}}, \code{\link{seqPtsOptNet}}, \code{\link{simPtsOptNet}} 
}
