\name{lalib}
\alias{lalib}
\docType{data}
\title{
Map of boundary enclosing La Libertad Research Center
}
\description{
Map of boundary enclosing La Libertad Research Center
}
\usage{data(lalib)}
\format{
  The format is:
Formal class 'SpatialPolygonsDataFrame' [package "sp"]
}
\details{
Map of boundary enclosing La Libertad Research Center. Spatial reference system: UTM 18N 
}
\source{
Santacruz, A. 2010. \emph{Design of optimal spatial sampling networks for the monitoring of soil organic carbon at La Libertad Research Center through the application of genetic algorithms}. M.Sc. Thesis. National University of Colombia, Bogota. 162 p. (In Spanish)
}
\references{
Santacruz, A., Rubiano, Y., Melo, C., 2014. \emph{Evolutionary optimization of spatial sampling networks designed for the monitoring of soil carbon}. In: Hartemink, A., McSweeney, K. (Eds.). \emph{Soil Carbon. Series: Progress in Soil Science}. (pp. 77-84). Springer. \href{https://link.springer.com/chapter/10.1007/978-3-319-04084-4_8}{[link]}

Santacruz, A., 2011. \emph{Evolutionary optimization of spatial sampling networks. An application of genetic algorithms and geostatistics for the monitoring of soil organic carbon}. Editorial Academica Espanola. 183 p. ISBN: 978-3-8454-9815-7 (In Spanish)
}
\examples{
data(lalib)
summary(lalib)
plot(lalib)
}
\keyword{datasets}
