\name{lalib}
\alias{lalib}
\docType{data}
\title{
Map of boundary enclosing La Libertad Research Center
}
\description{
Map of boundary enclosing La Libertad Research Center
}
\usage{data(lalib)}
\format{
  The format is:
Formal class 'SpatialPolygonsDataFrame' [package "sp"]
}
\details{
Map of boundary enclosing La Libertad Research Center. Spatial reference system: UTM 18N 
}
\source{
Santacruz, A. 2010. \emph{Design of optimal spatial sampling networks for the monitoring of soil organic carbon at La Libertad Research Center through the application of genetic algorithms}. M.Sc. Thesis. National University of Colombia, Bogotá. 162 p. (In Spanish)
}
\references{
Santacruz, A. 2011. \emph{Evolutionary optimization of spatial sampling networks. An application of genetic algorithms and geostatistics for the monitoring of soil organic carbon}. Editorial Académica Española. 183 p. ISBN: 978-3-8454-9815-7 (In Spanish) 
}
\examples{
data(lalib)
summary(lalib)
plot(lalib)
}
\keyword{datasets}
